Improve data section routing for tables to account for edge cases
