+++ packages/builder/src/helpers/urlStateSync.js
@@ -38,7 +38,7 @@ export const syncURLToState = options => {
   let cachedRedirect = get(routify.redirect)
   let cachedPage = get(routify.page)
   let previousParamsHash = null
-  let debug = false
+  let debug = true
   const log = (...params) => debug && console.log(...params)
 
   // Navigate to a certain URL
