+++ packages/builder/src/pages/builder/app/[application]/data/table/[tableId]/_layout.svelte
@@ -1,1 +1,20 @@
+<script>
+  import { syncURLToState } from "helpers/urlStateSync"
+  import { tables } from "stores/backend"
+  import * as routify from "@roxi/routify"
+  import { onDestroy } from "svelte"
+
+  const stopSyncing = syncURLToState({
+    urlParam: "tableId",
+    stateKey: "selectedTableId",
+    validate: id => $tables.list?.some(table => table._id === id),
+    update: tables.select,
+    fallbackUrl: "../",
+    store: tables,
+    routify,
+  })
+
+  onDestroy(stopSyncing)
+</script>
+
 <slot />
